Tests/LibWeb/Text/expected/HTML/PageSwapEvent-constructor.txt

PageSwapEvent is function
NavigationActivation is function
[object PageSwapEvent]
true
true
pageswap
false
false
null
null
